requirements.txt
requests==2.31.0
loguru==0.7.2
websockets==12.0
pyyaml==6.0.1